+++ src/wx.js
@@ -4,12 +4,59 @@ import {
     invokePageMethod
 } from './util.js';
 
+/**
+ * 如果跳转失败, 则尝试降级(degrade)到降级的方法(例如 `switchTab`)来完成跳转
+ * 
+ * @param {object} degradeFunctionName 降级的方法名
+ * @return {function} function(options)
+ */
+function failThenTryDegrade(degradeFunctionName) {
+    return function(options) {
+        var originalFail = options.fail;
+        var originalUrl = options.url;
+        function tryDegrade() {
+            var _options = extend({}, options);
+
+            // 调用过 `wx` 的跳转方法后, 会改写 `options.url` 参数
+            // 例如 `url` 原始值为 `/pages/home/index?a=1`
+            // 会被改写为 `pages/home/index.html?a=1`
+            _options.url = originalUrl;
+
+            // 如果降级的方法也调用失败了, 才触发失败回调
+            if (originalFail) {
+                _options.fail = originalFail;
+            } else {
+                // 删除用于降级的 fail 回调
+                delete _options.fail;
+            }
+
+            wx[degradeFunctionName](_options);
+        }
+
+        options.fail = function(result) {
+            // XXX 微信小程序没有提供错误码用于判定是想跳转到 tabbar 页面却用错了方法
+            // 例如: `navigateTo:fail can not navigateTo a tabbar page`
+            console.warn(result.errMsg);
+            if (originalUrl) {
+                // 标记尝试使用降级的方法来跳转页面
+                options._tryDegrade = degradeFunctionName;
+                tryDegrade();
+            }
+        };
+    };
+}
+
+var failThenTrySwitchTab = failThenTryDegrade('switchTab');
+var failThenTryRedirectTo = failThenTryDegrade('redirectTo');
+
 // 包装 wx 的方法
 
 /**
  * 封装原生的 navigateTo 方法
  * 
- * - 解决微信小程序只允许跳转 10 层路由的问题, 超过限制后自动变为 redirectTo 跳转页面
+ * - 扩展 `_urlParams` 用于在 URL 上追加参数
+ * - 调用失败时尝试降级为 `switchTab` 来完成跳转
+ * - 解决微信小程序只允许跳转 10 层路由的问题, 超过限制后自动变为 `redirectTo` 跳转页面
  * 
  * @param {object} options
  *                 options._urlParams {object} 需要追加到 URL 上的参数
@@ -20,6 +67,8 @@ export function navigateTo(options) {
     if (_options._urlParams) {
         _options.url = appendUrl(_options.url, _options._urlParams);
     }
+
+    failThenTrySwitchTab(_options);
 
     // 最多 10 次路由(首页本身也算一次)
     if (getCurrentPages().length < 10) {
@@ -32,6 +81,9 @@ export function navigateTo(options) {
 /**
  * 封装原生的 redirectTo 方法
  * 
+ * - 扩展 `_urlParams` 用于在 URL 上追加参数
+ * - 调用失败时尝试降级为 `switchTab` 来完成跳转
+ * 
  * @param {object} options
  *                 options._urlParams {object} 需要追加到 URL 上的参数
  */
@@ -42,7 +94,20 @@ export function redirectTo(options) {
         _options.url = appendUrl(_options.url, _options._urlParams);
     }
 
+    failThenTrySwitchTab(_options);
     return wx.redirectTo(_options);
+};
+
+/**
+ * 封装原生的 `switchTab` 方法
+ * 
+ * - 调用失败时尝试降级为 `redirectTo` 来完成跳转
+ * 
+ * @param {object} options
+ */
+export function switchTab(options) {
+    failThenTryRedirectTo(options);
+    return wx.switchTab(options);
 };
 
 /**